requirements.txt
kubernetes
pyyaml
dacite
prometheus_client